refactor: refactoring methods
