+++ src/UseHomareHandler.ts
@@ -1,8 +1,8 @@
 import { Context } from 'hono';
 import { DiscordNotification } from './types/DiscordNotification';
 import { YouTubeFeed } from './types/youtubeXmlInterface';
-import { parseYouTubeFeed, YouTubeFeedParseError } from './parseXml';
-import { sendDiscordNotification, DiscordNotificationSendError } from './sendNotify';
+import { useYouTubeFeed, YouTubeFeedParseError } from './UseYouTubeFeed';
+import { useDiscordNortification, DiscordNotificationSendError } from './UseDiscordNotification';
 
 export const useHomareHandler = () => {
   // Result type for parsing YouTube feed
@@ -12,6 +12,7 @@ export const useHomareHandler = () => {
 
   // Safe parsing function that returns Result type
   const tryParseYouTubeFeed = (contextBody: string): YouTubeFeedParseResult => {
+    const { parseYouTubeFeed } = useYouTubeFeed();
     try {
       return { success: true, data: parseYouTubeFeed(contextBody) };
     } catch (error) {
@@ -39,6 +40,7 @@ export const useHomareHandler = () => {
    * 痺れますね！
    */
   const postShibireMasuNeNotification = async (context: Context) => {
+    const { createDiscordNotification, sendDiscordNotification } = useDiscordNortification();
     // parse YouTube feed from context using Result pattern
     const contextBody: string = await context.req.text();
     const youTubeFeedParseResult: YouTubeFeedParseResult = tryParseYouTubeFeed(contextBody);
@@ -51,11 +53,7 @@ export const useHomareHandler = () => {
     const youTubeFeed: YouTubeFeed = youTubeFeedParseResult.data;
 
     // create discord notification from YouTube feed
-    const discordNotification: DiscordNotification = {
-      message: '新着動画だよ！（暖かみのあるbot）',
-      title: youTubeFeed.feed.entry.title,
-      url: youTubeFeed.feed.entry.link['@_href'],
-    };
+    const discordNotification: DiscordNotification = createDiscordNotification(youTubeFeed);
 
     // send discord notification
     const webhookUrl: string = (context.env as { DISCORD_WEBHOOK_URL: string }).DISCORD_WEBHOOK_URL;
